% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-olson.R
\name{abbreviate_olson}
\alias{abbreviate_olson}
\title{Abbreviate an Olson timezone name}
\usage{
abbreviate_olson(tz, width = 14L, dictionary = NULL)
}
\arguments{
\item{tz}{\code{character}, length-one, timezone-name to abbreviate}

\item{width}{\code{integer}, maximum number of characters}

\item{dictionary}{\code{character}, named vector: values are abbreviated
components, names are unabbreviated components}

\item{tz_compopnent}{\code{character}, timezone-name component to abbreviate}
}
\value{
\code{character}, abbreviated name
}
\description{
Abbreviate an Olson timezone name
}
\examples{
abbreviate_olson("America/Chicago")
abbreviate_olson("America/Chicago", width = 9)
abbreviate_olson("America/Chicago", dictionary = c(America = "USA"))
}
\keyword{internal}
