% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-data.R
\name{cf_data}
\alias{cf_data}
\alias{cf_data.logical}
\alias{cf_data.numeric}
\alias{cf_data.Date}
\alias{cf_data.POSIXct}
\alias{cf_data.character}
\title{Column data}
\usage{
cf_data(x, ...)

\method{cf_data}{logical}(x, ...)

\method{cf_data}{numeric}(x, ..., sigfig = 3)

\method{cf_data}{Date}(x, ...)

\method{cf_data}{POSIXct}(x, ...)

\method{cf_data}{character}(x, ...)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Other arguments passed to methods}

\item{sigfig}{Minimum number of significant figures to display. Numbers
larger than 1 will potentially show more signficiant figures than this
but they will be greyed out.}
}
\description{
Internal class for formatting the data part of a column.
}
