% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char.R
\name{char}
\alias{char}
\alias{set_char_opts}
\title{Format a character vector in a tibble}
\usage{
char(
  x,
  ...,
  min_chars = NULL,
  shorten = c("back", "front", "mid", "abbreviate")
)

set_char_opts(
  x,
  ...,
  min_chars = NULL,
  shorten = c("back", "front", "mid", "abbreviate")
)
}
\description{
These functions are reexported as \code{\link[tibble:char]{tibble::char()}} and \code{\link[tibble:char]{tibble::set_char_opts()}}.
}
\keyword{internal}
