% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldata.R
\name{coldata}
\alias{coldata}
\alias{coldata.logical}
\alias{coldata.numeric}
\alias{coldata.Date}
\alias{coldata.POSIXct}
\alias{coldata.character}
\title{Format a vector suitable for tabular display}
\usage{
coldata(x, ...)

\method{coldata}{logical}(x, ...)

\method{coldata}{numeric}(x, ..., sigfig = 3, sci_threshold = 15)

\method{coldata}{Date}(x, ...)

\method{coldata}{POSIXct}(x, ...)

\method{coldata}{character}(x, ..., width = NA)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Other arguments passed to methods}

\item{sigfig}{Minimum number of significant figures to display. Numbers
larger than 1 will potentially show more signficiant figures than this
but they will be greyed out.}

\item{sci_threshold}{If decimal display is wider than this threshold,
use scientific display instead.}

\item{width}{Preferred width of output}
}
\value{
A character vector with class `coldata` and
  `width` and `align` attributes.
}
\description{
Format a vector suitable for tabular display
}
\examples{
x <- 123456789 * (10 ^ c(-1, -3, -5, NA, -8, -10))
coldata(x)
coldata(-x)
coldata(runif(10))
coldata(rcauchy(20))

# Special values are highlighted
coldata(c(runif(5), NA, NaN, Inf, -Inf))

# Very wide ranges will be displayed in scientific format
coldata(c(1e10, 1e-10))
coldata(c(1e10, 1e-10), sci_threshold = Inf)

x <- c(FALSE, NA, FALSE, FALSE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)
coldata(x)

x <- c("This is string is rather long", NA, "?", "Short")
coldata(x)
coldata(x, width = 30)
coldata(x, width = 5)

date <- as.Date("2017-05-15")
coldata(date + c(1, NA, 3:5))
coldata(as.POSIXct(date) + c(30, NA, 600, 3600, 86400))
}
