% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat}
\alias{colformat}
\alias{colformat.logical}
\alias{colformat.Date}
\alias{colformat.POSIXct}
\alias{colformat.character}
\title{Format a vector suitable for tabular display}
\usage{
colformat(x, ...)

\method{colformat}{logical}(x, ...)

\method{colformat}{Date}(x, ...)

\method{colformat}{POSIXct}(x, ...)

\method{colformat}{character}(x, ..., width = NA)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Other arguments passed to methods}

\item{width}{Preferred width of output}
}
\value{
A character vector with class `colformat` and
  `width` and `align` attributes.
}
\description{
Format a vector suitable for tabular display
}
\examples{

x <- c(FALSE, NA, FALSE, FALSE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)
colformat(x)

x <- c("This is string is rather long", NA, "?", "Short")
colformat(x)
colformat(x, width = 30)
colformat(x, width = 5)

date <- as.Date("2017-05-15")
colformat(date + c(1, NA, 3:5))
colformat(as.POSIXct(date) + c(30, NA, 600, 3600, 86400))
}
