% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfig.R
\name{decimal_format}
\alias{decimal_format}
\title{Create a decimal table}
\usage{
decimal_format(x, sigfig = 3)
}
\value{
A tibble with four columns:
* `neg`: negative sign or space, if needed
* `lhs`: whole number
* `dec`: decimal point, if needed
* `rhs`: remainder of number
}
\description{
This is a key building block of number formatting. It carefully uses
colour and alignment to make it as easy as possible to compare numbers.
}
\examples{
x <- 123456789 * (10 ^ c(1, -3, -5, NA, -8, -10, -15))
decimal_format(x, 3)

x <- x * sample(c(-1, 1), length(x), rep = TRUE)
decimal_format(x, 3)

decimal_format(c(Inf, -Inf, NA, NaN), 3)
decimal_format(c(1e10, 1e-10), 3)
}
\keyword{internal}
