% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{is_vector_s3}
\title{Deprecated functions}
\usage{
is_vector_s3(x)
}
\description{
\code{is_vector_s3()} forwards to \code{\link[vctrs:vec_assert]{vctrs::vec_is()}}.
}
\section{Lifecycle}{

\code{is_vector_s3()} is soft-deprecated and no longer generic,
use \code{\link[vctrs:vec_assert]{vctrs::vec_is()}} instead.
}

\keyword{internal}
