% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{expect_colformat_output}
\alias{expect_colformat_output}
\alias{add_special}
\title{Test helpers}
\usage{
expect_colformat_output(x, ..., filename, xp = add_special(x),
  xf = colformat(xp, ...), crayon = TRUE)

add_special(x)
}
\arguments{
\item{x}{An object to be formatted.}

\item{...}{Passed on to \code{\link[=colformat]{colformat()}} if \code{xf} is left at its default.}

\item{filename}{File name that contains the desired output.}

\item{xp}{Pass a value here instead of \code{x} if you want to omit appending
\code{NA} and \code{Inf} values.}

\item{xf}{Pass the result of a \code{\link[=colformat]{colformat()}} call here for full control.}

\item{crayon}{Color the output?}
}
\description{
Unexported functions for packages that implement their own colformat.
}
