% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{expect_known_display}
\alias{expect_known_display}
\title{Test helpers}
\usage{
expect_known_display(object, file, ..., width = 80L, crayon = TRUE)
}
\arguments{
\item{object}{An object to check.}

\item{file}{File path where known value/output will be stored.}

\item{...}{Unused.}

\item{width}{The width of the output.}

\item{crayon}{Color the output?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

Expectation for packages that implement a data type with pillar support.
Allows storing the desired result in a file,
and comparing the output with the file contents.
Note that this expectation sets options that affect the formatting of the
pillar, see examples for usage.
}
\section{Life cycle}{

This function is deprecated in favor of \code{\link[testthat:expect_snapshot]{testthat::expect_snapshot()}}.
If your package uses the third edition of testthat, do not use this function.
}

\keyword{internal}
