% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{expect_pillar_output}
\alias{expect_pillar_output}
\alias{add_special}
\title{Test helpers}
\usage{
expect_pillar_output(x, ..., filename, xp = add_special(x), xf = pillar(xp,
  ...), crayon = TRUE)

add_special(x)
}
\arguments{
\item{x}{An object to be formatted.}

\item{...}{Passed on to \code{\link[=pillar]{pillar()}} if \code{xf} is left at its default.}

\item{filename}{File name that contains the desired output.}

\item{xp}{Pass a value here instead of \code{x} if you want to omit appending
\code{NA} and \code{Inf} values.}

\item{xf}{Pass the result of a \code{\link[=pillar]{pillar()}} call here for full control.}

\item{crayon}{Color the output?}
}
\description{
Helper functions for packages that implement their own pillar.
\code{expect_pillar_output()} is an expectation that allows storing the
desired result in a file, and comparing the output with the file contents.

\code{add_special()} is not exported, and used only for initializing default
values to \code{expect_pillar_output()}.
}
