% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific.R
\name{format_scentific}
\alias{format_scentific}
\title{Format numbers in scientific notation}
\usage{
format_scentific(x, sigfig = 3)
}
\arguments{
\item{x}{A numeric vector}

\item{sigfig}{Number of signficiant figures}
}
\description{
Uses colour, careful alignment, and superscripts to display numbers
in scientific notation.
}
\examples{
x <- c(runif(10) * 10 ^ (sample(-100:100, 5)), NA, Inf, NaN)
format_scentific(x)
}
