% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{get_extent}
\alias{get_extent}
\alias{get_max_extent}
\title{Calculate display width}
\usage{
get_extent(x)

get_max_extent(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
\code{get_extent()} calculates the display width for each string in a character vector.

\code{get_max_extent()} calculates the maximum display width of all strings in a character vector,
zero for empty vectors.
}
\examples{
get_extent(c("abc", "de"))
get_extent("\u904b\u6c23")
get_max_extent(c("abc", "de"))
}
