% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multicolformat}
\alias{multicolformat}
\title{Format multiple vectors in a tabular display}
\usage{
multicolformat(x, width = NULL, ...)
}
\arguments{
\item{x}{A list of vectors to format}

\item{width}{Default width of the entire output, optional}

\item{...}{Ignored}
}
\description{
Format multiple vectors in a tabular display
}
