% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl_pillar_box.R
\name{new_pillar_box}
\alias{new_pillar_box}
\alias{pillar_box}
\title{TBD}
\usage{
new_pillar_box(x, ..., width, min_width = NULL)

pillar_box(x)
}
\arguments{
\item{x}{TBD}

\item{...}{These dots are for future extensions and must be empty.}

\item{width, min_width}{TBD}
}
\description{
TBD
}
