% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\name{new_pillar_type}
\alias{new_pillar_type}
\title{Prepare a column type for formatting}
\usage{
new_pillar_type(x, ...)
}
\arguments{
\item{x}{A vector for which the type is to be retrieved.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Calls \code{\link[=type_sum]{type_sum()}} to format the type.
 Call \code{\link[=format]{format()}} on the result to render column types.
}
\examples{
format(new_pillar_type("a"))
format(new_pillar_type(factor("a")))
}
