% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ornament.R
\name{new_vertical}
\alias{new_vertical}
\alias{print.pillar_vertical}
\title{Vertical text}
\usage{
new_vertical(x, ..., subclass = NULL)

\method{print}{pillar_vertical}(x, ...)
}
\description{
This class implements a \code{\link[=print]{print()}} method which outputs the contents
(usually a character vector) to the console using \code{\link[=cat]{cat()}},
without any extra marks and ignoring width or length restrictions.
}
\examples{
x <- new_vertical(letters[1:3])
str(x)
x
}
