% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num.R
\name{num}
\alias{num}
\alias{set_num_opts}
\title{Format a numeric vector in a tibble}
\usage{
num(
  x,
  ...,
  sigfig = NULL,
  digits = NULL,
  label = NULL,
  scale = NULL,
  notation = c("fit", "dec", "sci", "eng", "si"),
  fixed_exponent = NULL,
  extra_sigfig = NULL
)

set_num_opts(
  x,
  ...,
  sigfig = NULL,
  digits = NULL,
  label = NULL,
  scale = NULL,
  notation = c("fit", "dec", "sci", "eng", "si"),
  fixed_exponent = NULL,
  extra_sigfig = NULL
)
}
\value{
A \code{pillar_num} vector for \code{num()}, or \code{x} with pillar attributes
set for \code{set_num_opts()}.
}
\description{
These functions are reexported as \code{\link[tibble:num]{tibble::num()}} and \code{\link[tibble:set_num_opts]{tibble::set_num_opts()}}.
}
\keyword{internal}
