% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width.R
\name{set_width}
\alias{set_width}
\alias{set_min_width}
\title{Set width and minimum width information for an object}
\usage{
set_width(x, width)

set_min_width(x, min_width)
}
\arguments{
\item{x}{Input to which assign a width or minimum width}

\item{width, min_width}{The new width}
}
\description{
Returns the input with updated \code{width} or \code{min_width} attributes.
}
