% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark-bar.R
\name{spark_bar}
\alias{spark_bar}
\title{Draw a sparkline bar graph with unicode block characters}
\usage{
spark_bar(x)
}
\arguments{
\item{x}{A numeric vector between 0 and 1}
}
\description{
Draw a sparkline bar graph with unicode block characters
}
\examples{
x <- seq(0, 1, length = 15)
spark_bar(x)
spark_bar(sample(x))

spark_bar(c(0, NA, 0.5, NA, 1))
}
