% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style_subtle_num}
\alias{style_subtle_num}
\title{\code{style_subtle_num()} is affected by the \code{pillar.subtle_num} option, which is
\code{FALSE} by default.}
\usage{
style_subtle_num(x)
}
\description{
\code{style_subtle_num()} is affected by the \code{pillar.subtle_num} option, which is
\code{FALSE} by default.
}
\examples{
style_subtle_num(0.01)
}
