% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-format-setup.R
\name{tbl_nrow}
\alias{tbl_nrow}
\title{Number of rows in a tbl object}
\usage{
tbl_nrow(x, ...)
}
\arguments{
\item{x}{A tbl object.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
This generic will be called by \code{\link[=tbl_format_setup]{tbl_format_setup()}} to determine the number of rows in a tbl object.
}
