% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-olson.R
\name{abbreviate_olson}
\alias{abbreviate_olson}
\title{Abbreviate an Olson timezone name}
\usage{
abbreviate_olson(tz, minwidth = 14L, dictionary = NULL, ...)
}
\arguments{
\item{tz}{\code{character}, timezone names to abbreviate}

\item{minwidth}{\code{integer}, desired length of abbreviation}

\item{dictionary}{\code{character}, named vector to override default dictionary:
values are abbreviated components, names are unabbreviated components}

\item{...}{other arguments passed onto \code{abbreviate()}}
}
\value{
\code{character}, abbreviated timezone names
}
\description{
Within this package, a date-time column is given a heading according to its
Olson-timezone name. Because some timezone names are too long to fit in the
space allocated for the header, an abbreviation function is needed.
}
\details{
The \code{minwidth} argument is used to specify the width to which you wish to
abbreviate a timezone name. This argument works like the \code{minwidth} argument
to \code{abbreviate()}: if the abbreviation returned has this number of
characters or fewer, the abbreviation is unique. If a longer abbreviation is
returned, this is an indication that \code{abbrevaiate()} was not able to make it
shorter and be unique.

Some of the common components of timezone names are given default
abbreviations, i.e. \code{"America"} becomes \code{"Amer"}. Use the \code{dictionary}
argument to override defaults.
}
\examples{
abbreviate_olson("America/Chicago")
abbreviate_olson("America/Chicago", minwidth = 20)
abbreviate_olson("America/Chicago", dictionary = c(America = "USA"))
abbreviate_olson(OlsonNames())
}
\seealso{
OlsonNames()
}
\keyword{internal}
