% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-olson.R
\name{abbreviate_olson}
\alias{abbreviate_olson}
\title{Abbreviate an Olson timezone name}
\usage{
abbreviate_olson(tz, width = 14L, consistent = TRUE, dictionary = NULL)
}
\arguments{
\item{tz}{\code{character}, length-one, timezone-name to abbreviate}

\item{width}{\code{integer}, maximum number of characters}

\item{consistent}{\code{logical}, always abbvreviate first and second
level names, i.e. \code{Africa/Bissau} is abbreviated to \code{Afr/Bissau} regardless
of available space}

\item{dictionary}{\code{character}, named vector: values are abbreviated
components, names are unabbreviated components}
}
\value{
\code{character}, abbreviated name
}
\description{
Abbreviate an Olson timezone name
}
\examples{
abbreviate_olson("America/Chicago")
abbreviate_olson("America/Chicago", minwidth = 20)
abbreviate_olson("America/Chicago", dictionary = c(America = "USA"))
abbreviate_olson(OlsonNames())
}
\seealso{
OlsonNames()
}
\keyword{internal}
