% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldata.R
\name{col_data}
\alias{col_data}
\alias{col_data.logical}
\alias{col_data.numeric}
\alias{col_data.Date}
\alias{col_data.POSIXct}
\alias{col_data.character}
\title{Column data}
\usage{
col_data(x, ...)

\method{col_data}{logical}(x, ...)

\method{col_data}{numeric}(x, ..., sigfig = 3, sci_threshold = 15)

\method{col_data}{Date}(x, ...)

\method{col_data}{POSIXct}(x, ...)

\method{col_data}{character}(x, ..., width = NA)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Other arguments passed to methods}

\item{sigfig}{Minimum number of significant figures to display. Numbers
larger than 1 will potentially show more signficiant figures than this
but they will be greyed out.}

\item{sci_threshold}{If decimal display is wider than this threshold,
use scientific display instead.}

\item{width}{Preferred width of output}
}
\description{
Internal class for formatting the data part of a column.
}
\examples{
x <- 123456789 * (10 ^ c(-1, -3, -5, NA, -8, -10))
col_data(x)
col_data(-x)
col_data(runif(10))
col_data(rcauchy(20))

# Special values are highlighted
col_data(c(runif(5), NA, NaN, Inf, -Inf))

# Very wide ranges will be displayed in scientific format
col_data(c(1e10, 1e-10))
col_data(c(1e10, 1e-10), sci_threshold = Inf)

x <- c(FALSE, NA, FALSE, FALSE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)
col_data(x)

x <- c("This is string is rather long", NA, "?", "Short")
col_data(x)
col_data(x, width = 30)
col_data(x, width = 5)

date <- as.Date("2017-05-15")
col_data(date + c(1, NA, 3:5))
col_data(as.POSIXct(date) + c(30, NA, 600, 3600, 86400))
}
