% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat}
\alias{colformat}
\title{Format a vector suitable for tabular display}
\usage{
colformat(x, ..., width = NA)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Other arguments passed to methods}

\item{width}{Preferred width of output}
}
\value{
A character vector where every element is the same width.
  Should have `colformat` for better default print method.
}
\description{
Format a vector suitable for tabular display
}
