% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat}
\alias{colformat}
\alias{colformat.logical}
\alias{colformat.numeric}
\alias{colformat.Date}
\alias{colformat.POSIXct}
\alias{colformat.character}
\title{Format a vector suitable for tabular display}
\usage{
colformat(x, ...)

\method{colformat}{logical}(x, ...)

\method{colformat}{numeric}(x, ..., sigfig = 3)

\method{colformat}{Date}(x, ...)

\method{colformat}{POSIXct}(x, ...)

\method{colformat}{character}(x, ..., width = NA)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Other arguments passed to methods}

\item{sigfig}{Minimum number of significant figures to display. Numbers
larger than 1 will potentially show more signficiant figures than this
but they will be greyed out.}

\item{width}{Preferred width of output}
}
\value{
A character vector with class `colformat` and
  `width` and `align` attributes.
}
\description{
Format a vector suitable for tabular display
}
\examples{
x <- 123456789 * (10 ^ c(1, -3, -5, NA, -8, -10, -15))
colformat(x)
colformat(-x)
colformat(runif(10))
colformat(rcauchy(10))
colformat(c(1, 0.5, 1e-10, NA, NaN, Inf, -Inf))

x <- c(FALSE, NA, FALSE, FALSE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)
colformat(x)

x <- c("This is string is rather long", NA, "?", "Short")
colformat(x)
colformat(x, width = 30)
colformat(x, width = 5)

date <- as.Date("2017-05-15")
colformat(date + c(1, NA, 3:5))
colformat(as.POSIXct(date) + c(30, NA, 600, 3600, 86400))
}
