% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat}
\alias{colformat}
\title{Format a vector suitable for tabular display}
\usage{
colformat(x, ..., width = NA)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Other arguments passed to methods}

\item{width}{Preferred width of output}
}
\value{
A character vector with class `colformat` and
  `width` and `align` attributes.
}
\description{
Format a vector suitable for tabular display
}
\examples{

x <- sample(c(TRUE, FALSE, NA), 10, replace = TRUE, prob = c(0.4, 0.4, 0.2))
colformat(x)

x <- c("This is string is rather long", NA, "?", "Short")
colformat(x)
colformat(x, width = 30)
colformat(x, width = 5)
}
