% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colformat.R
\name{colformat}
\alias{colformat}
\title{Format a vector suitable for tabular display}
\usage{
colformat(x, title = NULL, width = NULL, ...)
}
\arguments{
\item{x}{A vector to format}

\item{title}{An optional title for the column}

\item{width}{Default width, optional}

\item{...}{Other arguments passed to methods}
}
\description{
Format a vector suitable for tabular display
}
\examples{
x <- 123456789 * (10 ^ c(-1, -3, -5, NA, -8, -10))
colformat(x)
colformat(-x)
colformat(runif(10))
colformat(rcauchy(20))

# Special values are highlighted
colformat(c(runif(5), NA, NaN, Inf, -Inf))

# Very wide ranges will be displayed in scientific format
colformat(c(1e10, 1e-10), width = 20)
colformat(c(1e10, 1e-10))

x <- c(FALSE, NA, FALSE, FALSE, TRUE, FALSE, FALSE, TRUE, FALSE, TRUE)
colformat(x)

x <- c("This is string is rather long", NA, "?", "Short")
colformat(x)
colformat(x, width = 30)
colformat(x, width = 5)

date <- as.Date("2017-05-15")
colformat(date + c(1, NA, 3:5))
colformat(as.POSIXct(date) + c(30, NA, 600, 3600, 86400))
}
