% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{colonnade}
\title{Distributing pillars over multiple tiers}
\description{
Documentation generated from inline code comments.
}
\details{
For fitting pillars in one or more tiers, first a check is made
if all pillars fit with their maximum width (e.g.,
\code{option(tibble.width = Inf)} or narrow colonnade).
If yes, this is the resulting fit, no more work needs to be done.
Otherwise, if the maximum width is too wide, the same test
is carried out with the minimum width.
If this is still too wide, this is the resulting fit.
Otherwise, some tiers from the start
will contain pillars with their maximum width,
one tier will contain some pillars with maximum and some with minimum width,
and the remaining tiers contain pillars with their minimum width only.

For this, we compute a "reverse minimum assignment".

We determine the cut point where minimum and maximum assignment
agree.
The following strategy is applied:
\enumerate{
\item First, we determine the tier in which the cut point lies.
This is the first instance of a column that ends up in the same tier
for both minimum and maximum assignment.
\item A set of candidate cut points is derived.
\item We consult the column offsets. The last column where the minimum assignment
has a greater or equal offset than the maximum assignment is our latest
cut point.
If no such column exists, the cut point is the column just before our
first candidate.
\item Finally, we combine maximum and minimum reverse fits at the cut point.
We don't need to redistribute anything here.
}

Fitting pillars into tiers is very similar to a word-wrapping algorithm.
In a loop, new tiers are opened if the current tier overflows.
If a column is too wide to fit a single tier, it will never be
displayed, and the colonnade will be truncated there.
This case should never occur with reasonable display widths larger than
30 characters.
Truncation also happens if all available tiers are filled.

The remaining space is distributed from left to right.
Each column gains space proportional to the fraction of missing and
remaining space,
rounded down.
Any space remaining after rounding is distributed from left to right,
one space per column.
}
\keyword{internal}
