% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/testthat.R
\name{deprecated}
\alias{deprecated}
\alias{is_vector_s3}
\alias{expect_known_display}
\title{Deprecated functions}
\usage{
is_vector_s3(x)

expect_known_display(object, file, ..., width = 80L, crayon = TRUE)
}
\arguments{
\item{object}{An object to check.}

\item{file}{File path where known value/output will be stored.}

\item{...}{Unused.}

\item{width}{The width of the output.}

\item{crayon}{Color the output?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Use \code{\link[vctrs:vec_assert]{vctrs::vec_is()}} instead of \code{is_vector_s3()}.

Use \code{\link[testthat:expect_snapshot]{testthat::expect_snapshot()}} instead of \code{expect_known_display()}.
}
\keyword{internal}
