% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{is_vector_s3}
\title{Deprecated functions}
\usage{
is_vector_s3(x)
}
\description{
\code{is_vector_s3()} returns \code{TRUE} if an object represents an S3 vector.
The default implementation forwards to \code{\link[vctrs:vec_assert]{vctrs::vec_is()}}.
}
\section{Lifecycle}{

\code{is_vector_s3()} is soft-deprecated, because \code{\link[vctrs:vec_assert]{vctrs::vec_is()}}
provides the same functionality based on a robust framework.
The generic will remain available for a while but is no longer used.
Users and implementers are encouraged to ensure that \code{vctrs::vec_is()}
is \code{TRUE} for their class.
}

\keyword{internal}
