% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim_desc}
\alias{dim_desc}
\title{Format dimensions}
\usage{
dim_desc(x)
}
\arguments{
\item{x}{The object to format the dimensions for}
}
\value{
A string describing the dimensions, e.g. \code{"10"} for a vector
or \code{"32 x 11"} for a data frame.
}
\description{
Multi-dimensional objects are formatted as \verb{a x b x ...}, for vectors the
length is returned.
}
\examples{
dim_desc(1:10)
dim_desc(Titanic)
}
