% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.R
\name{dim_desc}
\alias{dim_desc}
\title{Format dimensions}
\usage{
dim_desc(x)
}
\arguments{
\item{x}{The object to format the dimensions for}
}
\description{
Multi-dimensional objects are formatted as `a x b x ...`, for vectors the
length is returned.
}
\examples{
dim_desc(1:10)
dim_desc(Titanic)
}
