% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat.R
\name{expect_known_display}
\alias{expect_known_display}
\title{Test helpers}
\usage{
expect_known_display(object, file, ..., crayon = TRUE, width = 80L)
}
\arguments{
\item{object}{object to test}

\item{file}{Path to a "golden" text file that contains the desired output.}

\item{...}{Unused.}

\item{crayon}{Color the output?}

\item{width}{Output width.}
}
\description{
Helper functions for packages that implement their own pillar.
\code{expect_pillar_output()} is an expectation that allows storing the
desired result in a file, and comparing the output with the file contents.
}
