% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{extra_cols}
\alias{extra_cols}
\title{Retrieve information about columns that didn't fit the available width}
\usage{
extra_cols(x, ...)
}
\arguments{
\item{x}{The result of \code{\link[=squeeze]{squeeze()}} on a \link{colonnade} object}

\item{...}{Unused}
}
\description{
Formatting a \link{colonnade} object may lead to some columns being omitted
due to width restrictions. This method returns a character vector that
describes each of the omitted columns.
}
\examples{
extra_cols(squeeze(colonnade(list(a = 1:3, b = 4:6), width = 8)))
}
