% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{extra_cols}
\alias{extra_cols}
\title{Retrieve information about columns that didn't fit the available width}
\usage{
extra_cols(x, ...)
}
\arguments{
\item{x}{The result of [format()] on a [multicolformat] object}

\item{...}{Unused}
}
\description{
Formatting a [multicolformat] object may lead to some columns being omitted
due to width restrictions. This method returns a character vector that
describes each of the omitted columns.
}
