% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfig.R
\name{format_decimal}
\alias{format_decimal}
\title{Format numbers in decimal notation}
\usage{
format_decimal(x, sigfig = 3, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{sigfig}{Number of signficiant figures to display.}

\item{...}{Ignored}
}
\value{
A list with at least the following elements:
\itemize{
\item \code{neg}: negative sign or space, if needed
\item \code{lhs}: whole number
\item \code{dec}: decimal point, if needed
\item \code{rhs}: remainder of number
}
}
\description{
This formatting system is designed to make it as easy as possible to
compare columns of numbers. Significant digits are coloured black or red
(for positive and negative numbers) and non-significant digits are coloured
in paler gray.
}
