% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scientific.R
\name{format_scientific}
\alias{format_scientific}
\title{Format numbers in scientific notation}
\usage{
format_scientific(x, sigfig = 3, superscript = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{sigfig}{Number of significant figures to display.}

\item{superscript}{If \code{TRUE}, will use superscript numbers in exponent.}
}
\description{
Uses colour, careful alignment, and superscripts to display numbers
in scientific notation.
}
\examples{
format_scientific(1.5:3.5)
format_scientific(1e9)
}
\seealso{
\code{\link[=format_decimal]{format_decimal()}}
}
