% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-format.R
\name{format_tbl}
\alias{format_tbl}
\alias{print.tbl}
\alias{format.tbl}
\title{Formatting of tbl objects}
\usage{
\method{print}{tbl}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{format}{tbl}(
  x,
  width = NULL,
  ...,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)
}
\description{
See \link[tibble:formatting]{tibble::formatting} for details.
}
\keyword{internal}
