% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-format.R
\name{format_tbl}
\alias{format_tbl}
\alias{print.tbl}
\alias{format.tbl}
\title{Formatting of tbl objects}
\usage{
\method{print}{tbl}(x, width = NULL, ..., n = NULL, n_extra = NULL)

\method{format}{tbl}(x, width = NULL, ..., n = NULL, n_extra = NULL)
}
\arguments{
\item{x}{Object to format or print.}

\item{width}{Width of text output to generate. This defaults to \code{NULL}, which
means use \code{getOption("tibble.width")} or (if also \code{NULL})
\code{getOption("width")}; the latter displays only the columns that fit on one
screen. You can also set \code{options(tibble.width = Inf)} to override this
default and always print all columns.}

\item{...}{Passed on to \code{\link[=tbl_format_setup]{tbl_format_setup()}} from \code{print()} and \code{format()},
must be empty for other methods.}

\item{n}{Number of rows to show. If \code{NULL}, the default, will print all rows
if less than option \code{tibble.print_max}. Otherwise, will print
\code{tibble.print_min} rows.}

\item{n_extra}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL}, the default,
will print information about at most \code{tibble.max_extra_cols} extra columns.}
}
\description{
These functions and methods are responsible for printing objects
of the \code{"tbl"} class, which includes \link[tibble:tibble]{tibble}s and dbplyr lazy
data frames.
See \link[tibble:formatting]{tibble::formatting} for user level documentation,
and \code{vignette("customization")} for details.
}
\examples{
print(vctrs::new_data_frame(list(a = 1), class = "tbl"))
}
\seealso{
\itemize{
\item \code{\link[=tbl_format_setup]{tbl_format_setup()}} for preparing an object for formatting
}
}
