% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multicolformat}
\alias{multicolformat}
\alias{squeeze}
\title{Format multiple vectors in a tabular display}
\usage{
multicolformat(x, has_row_id = TRUE, width = NULL, ...)

squeeze(x, width = NULL, ...)
}
\arguments{
\item{x}{A list of vectors to format}

\item{has_row_id}{Include a column indicating row IDs? Pass `"*"` to mark
the row ID column with a star.}

\item{width}{Default width of the entire output, optional}

\item{...}{Ignored}
}
\description{
The vectors are formatted to fit horizontally into a user-supplied number of
characters per row.

The `squeeze()` function is called by [format()]  and [print()] and usually
doesn't need to be called manually.
It returns an object suitable for printing and formatting at a fixed width
with additional information about omitted columns.
}
\details{
In a first pass, for each column it is decided if it is hidden, shown with
its minimum width or shown with its maximum width.
Remaining space is then distributed proportionally to columns that do not
use their desired width.

For computing the column widths, two cases are distinguished:
1. When taking the minimum width for each column (plus one inter-column
   space), at least one column does not fit.
   In this case, the minimum width is assigned to all columns that do fit,
   the non-fitting columns are stripped.
1. All columns fit with their minimum width. In this case, starting at
   the leftmost column, the maximum width is allocated to the columns
   until all available space is used.

The remaining space is distributed from left to right.
Each column gains space proportional to the fraction of missing and
remaining space,
rounded down.
Any space remaining after rounding is distributed from left to right,
one space per column.
}
