% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{new_pillar_title}
\alias{new_pillar_title}
\title{Prepare a column title for formatting}
\usage{
new_pillar_title(x, ...)
}
\arguments{
\item{x}{A character vector of column titles.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Call \code{\link[=format]{format()}} on the result to render column titles.
}
\examples{
format(new_pillar_title(names(iris)))
}
