% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-format-setup.R
\name{new_tbl_format_setup}
\alias{new_tbl_format_setup}
\title{Construct a setup object for formatting}
\usage{
new_tbl_format_setup(
  x,
  width,
  squeezed,
  rows_body,
  rows_missing,
  rows_total,
  extra_cols
)
}
\arguments{
\item{x}{The input object unchanged,}

\item{width}{The \code{width} argument unchanged,}

\item{squeezed}{TBD}

\item{rows_body}{The number of rows shown in the body,
even if the body does not have any columns.}

\item{rows_missing}{The number of rows not shown from the body,
\code{NA} if unknown.}

\item{rows_total}{The total number of rows in the data,
\code{NA} if unknown.}

\item{extra_cols}{TBD}
}
\description{
The object returned from the default method of \code{\link[=tbl_format_setup]{tbl_format_setup()}}
is a list with a \code{"class"} attribute and the elements described in the
"Parameters" section.

Named elements can be added to such objects without affecting the behavior.
Do not modify existing elements.
}
\keyword{internal}
