% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pillar-package.R
\docType{package}
\name{pillar-package}
\alias{pillar-package}
\title{pillar: Coloured Formatting for Columns}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Formats tabular data in columns or rows
using the full range of colours provided by modern terminals.
Provides various generics for making every aspect of the display customizable.
}
\examples{
\dontshow{if (rlang::is_installed("tibble")) withAutoprint(\{ # examplesIf}
pillar(1:3)
pillar(c(1, 2, 3))
pillar(factor(letters[1:3]), title = "letters")
tbl_format_setup(tibble::as_tibble(mtcars), width = 60)
\dontshow{\}) # examplesIf}
}
\seealso{
\itemize{
\item \code{\link[=pillar]{pillar()}} for formatting a single column,
\item \code{\link[=print.tbl]{print.tbl()}} for formatting data-frame-like objects,
\item \link{pillar_options} for a list of package options.
}
}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
