% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pillar-package.R
\docType{package}
\name{pillar-package}
\alias{pillar-package}
\title{pillar: Coloured Formatting for Columns}
\description{
Provides a 'pillar' generic designed for formatting columns
of data using the full range of colours provided by modern terminals.
}
\section{Options that control appearance}{

\itemize{
\item \code{pillar.bold}: Use bold font, e.g. for column headers? This currently
defaults to \code{FALSE}, because many terminal fonts have poor support for
bold fonts.
\item \code{pillar.subtle}: Use subtle style, e.g. for insignificant digits? Default:
\code{TRUE}.
}
}

\section{Options that control behaviour}{

\itemize{
\item \code{pillar.sigfig}: The number of significant digits that will be printed and
highlighted. Default: \code{3}. Set the \code{pillar.subtle} option to \code{FALSE} to
turn off highlighting of significant digits.
\item \code{pillar.colnames}: Controls the display of column names. The following
values are supported:
\itemize{
\item \code{"full"} (default): Always print full column names
\item \code{"truncate"}: Allow truncation of column names, may lead to duplicate
names in the header
}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/pillar}
  \item Report bugs at \url{https://github.com/r-lib/pillar/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Hadley Wickham
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
