% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{pillar_options}
\alias{pillar_options}
\title{Package options}
\usage{
pillar_options
}
\description{
Options that affect display of tibble-like output.
}
\details{
All options are available via the \code{pillar_options} list.
The elements of this list are combined getter/setter functions.
Calling a function without arguments returns the current value,
by providing an argument the current value is set and the old value
is returned, invisibly.
Setting \code{local = TRUE} enables the option for the duration of the
current stack frame via \code{\link[rlang:local_options]{rlang::local_options()}}.

These options can also be set via \code{\link[=options]{options()}} and queried via \code{\link[=getOption]{getOption()}}.
For this, add a \code{pillar.} prefix (the package name and a dot) to the option name.
Example: for an option \code{foo},
\code{pillar_options$foo(value)} is equivalent to
\code{options(pillar.foo = value)}.
}
\section{Options for the pillar package}{

\itemize{
\item \code{bold}: Use bold font, e.g. for column headers? This currently
defaults to \code{FALSE}, because many terminal fonts have poor support for
bold fonts.
\item \code{subtle}: Use subtle style, e.g. for row numbers and data types?
Default: \code{TRUE}.
\item \code{subtle_num}: Use subtle style for insignificant digits? Default:
\code{FALSE}, is also affected by the \code{subtle} option.
\item \code{neg}: Highlight negative numbers? Default: \code{TRUE}.
\item \code{sigfig}: The number of significant digits that will be printed and
highlighted, default: \code{3}. Set the \code{subtle} option to \code{FALSE} to
turn off highlighting of significant digits.
\item \code{min_title_chars}: The minimum number of characters for the column
title, default: \code{15}.  Column titles may be truncated up to that width to
save horizontal space. Set to \code{Inf} to turn off truncation of column
titles.
\item \code{min_chars}: The minimum number of characters wide to
display character columns, default: \code{3}.  Character columns may be
truncated up to that width to save horizontal space. Set to \code{Inf} to
turn off truncation of character columns.
\item \code{max_dec_width}: The maximum allowed width for decimal notation,
default 13.
}
}

\examples{
# Default setting:
pillar_options$sigfig()
pillar(1.234567)

# Change for the duration of the session:
old <- pillar_options$sigfig(6)
pillar(1.234567)

# Change back to the original value:
pillar_options$sigfig(old)
pillar(1.234567)

# Local scope:
local({
  pillar_options$sigfig(6, local = TRUE)
  pillar(1.234567)
})
pillar(1.234567)
}
\keyword{datasets}
