% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaft.R
\name{pillar_shaft}
\alias{pillar_shaft}
\alias{new_pillar_shaft}
\alias{pillar_shaft.logical}
\alias{pillar_shaft.numeric}
\alias{pillar_shaft.Date}
\alias{pillar_shaft.POSIXt}
\alias{pillar_shaft.character}
\alias{pillar_shaft.list}
\alias{pillar_shaft.default}
\title{Column data}
\usage{
pillar_shaft(x, ...)

new_pillar_shaft(x, ..., width = max(crayon::col_nchar(x, type = "width"),
  0L), align = "left", min_width = NULL, na_indent = 0L)

\method{pillar_shaft}{logical}(x, ...)

\method{pillar_shaft}{numeric}(x, ..., sigfig = 3)

\method{pillar_shaft}{Date}(x, ...)

\method{pillar_shaft}{POSIXt}(x, ...)

\method{pillar_shaft}{character}(x, ...)

\method{pillar_shaft}{list}(x, ...)

\method{pillar_shaft}{default}(x, ...)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Unused, for extensibility.}

\item{width}{The maximum column width.}

\item{align}{Alignment of the column.}

\item{min_width}{The minimum allowed column width, \code{width} if omitted.}

\item{na_indent}{Indention of \code{NA} values.}

\item{sigfig}{Minimum number of significant figures to display. Numbers
larger than 1 will potentially show more signficiant figures than this
but they will be greyed out.}
}
\description{
Internal class for formatting the data part of a column.
}
