% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shaft.R
\name{pillar_shaft}
\alias{pillar_shaft}
\alias{pillar_shaft.logical}
\alias{pillar_shaft.numeric}
\alias{pillar_shaft.Date}
\alias{pillar_shaft.POSIXt}
\alias{pillar_shaft.character}
\alias{pillar_shaft.list}
\alias{pillar_shaft.AsIs}
\alias{pillar_shaft.default}
\title{Column data}
\usage{
pillar_shaft(x, ...)

\method{pillar_shaft}{logical}(x, ...)

\method{pillar_shaft}{numeric}(x, ..., sigfig = 3)

\method{pillar_shaft}{Date}(x, ...)

\method{pillar_shaft}{POSIXt}(x, ...)

\method{pillar_shaft}{character}(x, ...)

\method{pillar_shaft}{list}(x, ...)

\method{pillar_shaft}{AsIs}(x, ...)

\method{pillar_shaft}{default}(x, ...)
}
\arguments{
\item{x}{A vector to format}

\item{...}{Unused, for extensibility.}

\item{sigfig}{Minimum number of significant figures to display. Numbers
larger than 1 will potentially show more signficiant figures than this
but they will be greyed out.}
}
\description{
Internal class for formatting the data for a column.
\code{pillar()} is a coercion method that must be implemented for your data type
to display it in a tibble.
}
\details{
The default method will currently display values coerced with
\code{\link[=as.character]{as.character()}} as a fallback, but you should not rely on this behavior.
}
