% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{squeeze}
\alias{squeeze}
\title{Squeeze a colonnade to a fixed width}
\usage{
squeeze(x, width = NULL, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}

The \code{squeeze()} function usually doesn't need to be called manually.
 It returns an object suitable for printing and formatting at a fixed width with additional information about omitted columns,
which can be retrieved via \code{\link[=extra_cols]{extra_cols()}}.
}
\keyword{internal}
