% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigfig.R, R/styles.R
\name{style_num}
\alias{style_num}
\alias{style_subtle}
\alias{style_na}
\alias{style_neg}
\title{Styling helpers}
\usage{
style_num(x, negative, significant = rep_along(x, TRUE))

style_subtle(x)

style_na(x)

style_neg(x)
}
\arguments{
\item{x}{The character vector to style.}

\item{negative, significant}{Logical vector the same length as \code{x} that
indicate if the values are negative and significant, respectively}
}
\description{
Functions that allow implementers of formatters for custom data types to
maintain a consistent style with the default data types.
}
