% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style_subtle}
\alias{style_subtle}
\alias{style_neg}
\title{Styling helpers}
\usage{
style_subtle(x)

style_neg(x)
}
\arguments{
\item{x}{The character vector to style.}
}
\description{
Functions that allow implementers of formatters for custom data types to
maintain a consistent style wit the default data types.
}
