% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-sum.R
\name{type_sum}
\alias{type_sum}
\title{Provide a succinct summary of an object}
\usage{
type_sum(x)
}
\arguments{
\item{x}{an object to summarise. Generally only methods of atomic vectors
and variants have been implemented.}
}
\description{
`type_sum()` gives a brief summary of object type. Objects that commonly
occur in a data frame should return a string with four or less characters.
}
